+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/exception/CodedException.kt
@@ -71,22 +71,39 @@ open class CodedException(
 inline fun <reified T : CodedException> errorCodeOf(): String =
   CodedException.inferCode(T::class.java)
 
-internal class IncompatibleArgTypeException(
-  argumentType: KClass<*>,
-  desiredType: KClass<*>,
-  cause: Throwable? = null
-) : CodedException(
-  message = "Argument type '$argumentType' is not compatible with expected type '$desiredType'.",
-  cause = cause
-)
+internal class IncompatibleArgTypeException : CodedException {
+  constructor(
+    argumentType: KClass<*>,
+    desiredType: KClass<*>,
+    cause: Throwable? = null
+  ) : super("Argument type '$argumentType' is not compatible with expected type '$desiredType'.", cause)
 
-internal class EnumNoSuchValueException(
-  enumType: KClass<Enum<*>>,
-  enumConstants: Array<out Enum<*>>,
-  value: Any?
-) : CodedException(
-  message = "'$value' is not present in ${enumType.simpleName} enum, it must be one of: ${enumConstants.joinToString(separator = ", ") { "'${it.name}'" }}"
-)
+  constructor(
+    argumentType: KClass<*>,
+    desiredType: Class<*>,
+    cause: Throwable? = null
+  ) : super("Argument type '$argumentType' is not compatible with expected type '${desiredType.simpleName}'.", cause)
+
+  constructor(
+    argumentType: Class<*>,
+    desiredType: Class<*>,
+    cause: Throwable? = null
+  ) : super("Argument type '${argumentType.simpleName}' is not compatible with expected type '${desiredType.simpleName}'.", cause)
+}
+
+internal class EnumNoSuchValueException : CodedException {
+  constructor(
+    enumType: KClass<Enum<*>>,
+    enumConstants: Array<out Enum<*>>,
+    value: Any?
+  ) : super("'$value' is not present in ${enumType.simpleName} enum, it must be one of: ${enumConstants.joinToString(separator = ", ") { "'${it.name}'" }}")
+
+  constructor(
+    enumType: Class<*>,
+    enumConstants: Array<out Enum<*>>,
+    value: Any?
+  ) : super("'$value' is not present in ${enumType.simpleName} enum, it must be one of: ${enumConstants.joinToString(separator = ", ") { "'${it.name}'" }}")
+}
 
 internal class MissingTypeConverter(
   forType: TypeDescriptor
@@ -276,11 +293,10 @@ internal class CollectionElementCastException private constructor(
 }
 
 @DoNotStrip
-class DynamicCastException(
-  type: KClass<*>
-) : CodedException(
-  message = "Could not cast dynamic value to '${type.qualifiedName}'."
-)
+class DynamicCastException : CodedException {
+  constructor(type: KClass<*>) : super("Could not cast dynamic value to '${type.qualifiedName}'.")
+  constructor(type: Class<*>) : super("Could not cast dynamic value to '${type.canonicalName}'.")
+}
 
 @DoNotStrip
 class JavaScriptEvaluateException(